src/userManagement/requirements.txt
Flask==3.1.0
Python-Dotenv==1.0.1
argon2-cffi==23.1.0
psycopg[binary,pool]==3.2.5